stupid modal chain animations
