+++ src/components/layoutManager.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useReducer, useEffect } from "react"
+import React, { useState, useReducer, useEffect, useRef } from "react"
 import { useSelector, useDispatch } from "react-redux"
 import styled from "styled-components"
 import { useSpring, useChain, animated, AnimatedValue } from "react-spring"
@@ -7,122 +7,9 @@ import Layout from "./layout"
 import Layout2 from "./layout2"
 import { setLayoutVersion, RootState } from "_reduxState"
 import { rhythm } from "utils/typography"
+import { FunButtonsModal } from "components/FunButtonsModal"
 
 import * as Colors from "consts/Colors"
-
-const XIcon = ({ fill }: { fill: AnimatedValue<any> }) => (
-  <animated.svg viewBox="0 0 24 24" s>
-    <animated.g>
-      <animated.path
-        fill={fill}
-        d="M13.414 12l5.793-5.793c.39-.39.39-1.023 0-1.414s-1.023-.39-1.414 0L12 10.586 6.207 4.793c-.39-.39-1.023-.39-1.414 0s-.39 1.023 0 1.414L10.586 12l-5.793 5.793c-.39.39-.39 1.023 0 1.414.195.195.45.293.707.293s.512-.098.707-.293L12 13.414l5.793 5.793c.195.195.45.293.707.293s.512-.098.707-.293c.39-.39.39-1.023 0-1.414L13.414 12z"
-      ></animated.path>
-    </animated.g>
-  </animated.svg>
-)
-const XIconContainer = styled.div`
-  border-radius: 100%;
-  position: absolute;
-  top: 0;
-  right: 0;
-  margin: 25px;
-  height: 25px;
-  width: 25px;
-
-  transition: background 200ms ease-in-out;
-  :hover {
-    background: lightgrey;
-  }
-`
-
-const ModalBackground = styled(animated.div)`
-  background-color: rgba(40, 40, 45, 0.8);
-  position: fixed;
-  top: 0;
-  left: 0;
-  min-width: 100vw;
-  min-height: 100vh;
-  z-index: 9998;
-`
-
-const Modal = styled(animated.div)`
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-
-  border-radius: 10px;
-  padding: 0px 50px 50px 50px;
-  height: 50%;
-  width: 25%;
-  background: white;
-  position: fixed;
-  /* top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%); */
-  /* margin-top: 50%; */
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-  margin: auto;
-
-  box-shadow: 0px 3px 5px -1px rgba(0, 0, 0, 0.2),
-    0px 6px 10px 0px rgba(0, 0, 0, 0.14), 0px 1px 18px 0px rgba(0, 0, 0, 0.12);
-  z-index: 9999;
-
-  @media (max-width: 769px) {
-    height: auto;
-    /* max-height: 500px; */
-    width: auto;
-    /* max-width: 300px; */
-
-    margin-top: 25%;
-    margin-right: 10%;
-    margin-bottom: 25%;
-    margin-left: 10%;
-  }
-`
-
-const ModalTitleText = styled(animated.h2)`
-  margin-bottom: 0;
-  @media (max-width: 350px) {
-    margin-top: 20%;
-  }
-`
-const Key = styled(animated.p)`
-  font-weight: 700;
-  font-size: 20px;
-  margin-bottom: 5px;
-`
-
-const ItemRowContainer = styled(animated.div)``
-const ItemRow = styled(animated.div)`
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-`
-const ButtonContainer = styled(animated.div)``
-const CheckboxContainer = styled(animated.div)`
-  text-align: center;
-`
-const Checkbox = styled(animated.input)``
-const Button = styled(animated.div)`
-  /* position: absolute; */
-  /* bottom: 0; */
-  border-radius: 10px;
-  border: 1px solid white;
-  background: grey;
-  text-align: center;
-  font-size: 20px;
-  font-weight: 500;
-  letter-spacing: 5px;
-  color: white;
-  transition: border 200ms ease-in-out;
-
-  :hover {
-    border: 1px solid black;
-  }
-`
 
 const LayoutManager = props => {
   const [showModal, toggleModal] = useReducer((s: boolean) => !s, false)
@@ -143,7 +30,7 @@ const LayoutManager = props => {
   /**
    * mount
    * ...if neverModalShowAgain is not in local storage
-   * 1. toggle modal
+   * 1. toggle modal TRUE
    * 2. increase opacity to 1
    */
   useEffect(() => {
@@ -160,10 +47,32 @@ const LayoutManager = props => {
     fill: isDarkMode ? Colors.silverLighter : Colors.blackDarker,
   }))
 
+  const resizeRef = useRef()
+  const { modalWidth, modalHeight } = useSpring({
+    ref: resizeRef,
+    from: { modalWidth: 0, modalHeight: 0 },
+    to: { modalWidth: shouldExit ? 0 : 25, modalHeight: shouldExit ? 0 : 50 },
+  })
+
+  const contentOpacityRef = useRef()
+  const { contentOpacity } = useSpring({
+    ref: contentOpacityRef,
+    from: { contentOpacity: 0 },
+    to: { contentOpacity: shouldExit ? 0 : 1 },
+  })
+
+  // useChain(open ? [springRef, transRef] : [transRef, springRef], [0, open ? 0.1 : 0.6])
+  useChain(
+    shouldExit
+      ? [contentOpacityRef, resizeRef]
+      : [resizeRef, contentOpacityRef],
+    shouldExit ? [0, 0] : [2.1, 3]
+  )
+
   /**
    * side effect of `shouldExit`
    * 1. decrease opacity to 0
-   * 2. then toggle modal
+   * 2. then toggle modal FALSE
    */
   useEffect(() => {
     if (shouldExit) {
@@ -208,59 +117,17 @@ const LayoutManager = props => {
   return (
     <>
       {showModal && (
-        <>
-          <Modal style={{ opacity, background: modalBackground }}>
-            <XIconContainer
-              onClick={() => {
-                setShouldExit(true)
-              }}
-            >
-              <XIcon fill={fill} />
-            </XIconContainer>
-            <ModalTitleText>Some Fun Buttons</ModalTitleText>
-
-            <ItemRowContainer>
-              <ItemRow>
-                <Key>T</Key> <span>Toggle Icon Trail</span>
-              </ItemRow>
-              <ItemRow>
-                <Key>S</Key> <span>Slow Mo</span>
-              </ItemRow>
-              <ItemRow>
-                <Key>F</Key> <span>???</span>
-              </ItemRow>
-              <ItemRow>
-                <Key>R</Key> <span>Reset</span>
-              </ItemRow>
-              <ItemRow>
-                <Key>D</Key> <span>Dark Mode</span>
-              </ItemRow>
-            </ItemRowContainer>
-
-            <ButtonContainer>
-              <CheckboxContainer>
-                <Checkbox
-                  type={"checkbox"}
-                  value={neverShowModalChecked}
-                  onChange={() => {
-                    toggleNeverShowModalChecked()
-                  }}
-                />{" "}
-                <small>Never show this again</small>
-              </CheckboxContainer>
-              <Button
-                onClick={() => {
-                  setShouldExit(true)
-                  neverShowModalChecked &&
-                    window.localStorage.setItem("neverModalShowAgain", "true")
-                }}
-              >
-                OK
-              </Button>
-            </ButtonContainer>
-          </Modal>
-          <ModalBackground style={{ opacity }} />
-        </>
+        <FunButtonsModal
+          animatedOpacity={opacity}
+          contentOpacity={contentOpacity}
+          animatedModalBackground={modalBackground}
+          animatedFill={fill}
+          modalWidth={modalWidth}
+          modalHeight={modalHeight}
+          setShouldExit={setShouldExit}
+          neverShowModalChecked={neverShowModalChecked}
+          toggleNeverShowModalChecked={toggleNeverShowModalChecked}
+        />
       )}
       {(() => {
         switch (layoutVersion) {
